backend/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
yfinance
pandas
numpy
pandas-ta
vaderSentiment
requests
python-dotenv
google-genai
gunicorn
uvicorn[standard]
